requirements.txt
# Core Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
audioread>=3.0.0
resampy>=0.4.0

# Deep Learning & ML Frameworks
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0

# Audio Enhancement & Processing
noisereduce>=2.0.0
pyloudnorm>=0.1.0
webrtcvad>=2.0.10
pyaudio>=0.2.13

# Speaker Diarization & Recognition
scikit-learn>=1.3.0

# ASR & Speech Processing
openai-whisper>=20231117
transformers>=4.36.0
tokenizers>=0.15.0
datasets>=2.15.0

# Web Framework & API
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Data Processing & Utilities
pandas>=2.1.0
tqdm>=4.66.0
Pillow>=10.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Configuration & Logging
PyYAML>=6.0.0
python-dotenv>=1.0.0
colorlog>=6.8.0

# HTTP Clients & Networking
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# System & OS
typing-extensions>=4.8.0